+++ pages-sections/projectMain/DibTravel.js
@@ -1,0 +1,22 @@
+import React from "react";
+
+import Taglines from "../dib-travel/Taglines";
+import Scooter from "../dib-travel/Scooter";
+import WebDesign from "../dib-travel/WebDesign";
+import SocialMedia from "../dib-travel/SocialMedia";
+import NewClients from "../dib-travel/NewClients";
+import PlaneAnimation from "../dib-travel/PlaneAnimation";
+function DibTravel() {
+  return (
+    <div>
+      <Taglines />
+      <Scooter />
+      <WebDesign />
+      <SocialMedia />
+      <NewClients />
+      <PlaneAnimation />
+    </div>
+  );
+}
+
+export default DibTravel;
