+++ pages-sections/projectMain/OwnNew.js
@@ -1,9 +1,9 @@
 import React from "react";
-import DesktopVideo from "../../pages-sections/ownnew/DesktopVideo";
-import Components from "../../pages-sections/ownnew/Components";
-import Mobile from "../../pages-sections/ownnew/Mobile";
-import Parallax from "../../pages-sections/ownnew/Parallax";
-import WebDevelopment from "../../pages-sections/ownnew/WebDevelopment";
+import DesktopVideo from "../ownnew/DesktopVideo";
+import Components from "../ownnew/Components";
+import Mobile from "../ownnew/Mobile";
+import Parallax from "../ownnew/Parallax";
+import WebDevelopment from "../ownnew/WebDevelopment";
 
 import { ParallaxProvider } from "react-scroll-parallax";
 function OwnNew() {
