--- pages/projects/DibTravel.js
@@ -1,22 +1,0 @@
-import React from "react";
-
-import Taglines from "../../pages-sections/dib-travel/Taglines";
-import Scooter from "../../pages-sections/dib-travel/Scooter";
-import WebDesign from "../../pages-sections/dib-travel/WebDesign";
-import SocialMedia from "../../pages-sections/dib-travel/SocialMedia";
-import NewClients from "../../pages-sections/dib-travel/NewClients";
-import PlaneAnimation from "../../pages-sections/dib-travel/PlaneAnimation";
-function DibTravel() {
-  return (
-    <div>
-      <Taglines />
-      <Scooter />
-      <WebDesign />
-      <SocialMedia />
-      <NewClients />
-      <PlaneAnimation />
-    </div>
-  );
-}
-
-export default DibTravel;
